almost working
The build of the java task runner does not work yet, but the python one
does
need to resolve the order and loading of the jar files.
the python files dont seem to be available to the java class at test
time,  I suppose will have to repackage them in the test dir.

more to come.

* README.md: